Acceptance tests

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
 
   def index
     @questions = Question.all
   end
 
   def show
+    @answer = question.answers.new
   end
 
   def new
@@ -17,8 +19,9 @@ class QuestionsController < ApplicationController
 
   def create
     @question = Question.new(question_params)
+    @question.author = current_user
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: "Your question successfully created."
     else
       render :new
     end
@@ -33,8 +36,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path
+    else
+      redirect_to question, notice: t('.destroy.errors.other')
+    end
   end
 
   private
